requirements.txt
Flask==3.1.0
gunicorn==22.0.0
autopep8==2.3.2
mysqlclient==2.1.1
python-dotenv==0.20.0
Flask-Mail==0.10.0
celery==5.4.0
redis==5.0.0
Flask-Caching==2.1.0
filedepot==0.8.0
cryptography==44.0.0
pytest==8.3.4
Flask-Login==0.6.3
Flask-WTF==1.2.0
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
